+++ components/documents/general-components/evaluation/final-evaluation-info.tsx
@@ -1,7 +1,7 @@
-import { ItemValue } from "../ui/ItemValue";
-import { NexusFormatDate } from "../ui/NexusFormatDate";
-import { SectionPaper } from "../paper/SectionPaper";
-import { SectionTitle } from "../paper/SectionTitle";
+import { ItemValue } from "../ui/item-value";
+import { NexusFormatDate } from "../ui/nexus-format-date";
+import { SectionPaper } from "../paper/section-paper";
+import { SectionTitle } from "../paper/section-title";
 
 type FinalEvaluationInfoType = {
   date: string;